--- PromoManagementApp/PromoManagementApp.sln
@@ -1,51 +1,0 @@
-﻿
-Microsoft Visual Studio Solution File, Format Version 12.00
-# Visual Studio Version 17
-VisualStudioVersion = 17.13.35806.99 d17.13
-MinimumVisualStudioVersion = 10.0.40219.1
-Project("{2150E333-8FDC-42A3-9474-1A3956D46DE8}") = "src", "src", "{02EA681E-C7D8-13C7-8484-4AC65E1B71E8}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PromoManagemntApp.Domain", "PromoManagemntApp.Domain\PromoManagemntApp.Domain.csproj", "{1F3838A6-3556-4665-9EDF-D2C43FAFEA2A}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PromoManagemntApp.Application", "PromoManagemntApp.Application\PromoManagemntApp.Application.csproj", "{1D76A876-71F7-4DB4-A8DB-B08D172A7F0C}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PromoManagemntApp.Infrastructure", "PromoManagemntApp.Infrastructure\PromoManagemntApp.Infrastructure.csproj", "{E5B1B762-2478-4CAC-85E7-B18DBD05E361}"
-EndProject
-Project("{FAE04EC0-301F-11D3-BF4B-00C04F79EFBC}") = "PromoManagemntApp.Presentation", "PromoManagemntApp.Presentation\PromoManagemntApp.Presentation.csproj", "{9C589DDD-0330-42D4-AAFE-CA6779BB87F8}"
-EndProject
-Global
-	GlobalSection(SolutionConfigurationPlatforms) = preSolution
-		Debug|Any CPU = Debug|Any CPU
-		Release|Any CPU = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(ProjectConfigurationPlatforms) = postSolution
-		{1F3838A6-3556-4665-9EDF-D2C43FAFEA2A}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1F3838A6-3556-4665-9EDF-D2C43FAFEA2A}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1F3838A6-3556-4665-9EDF-D2C43FAFEA2A}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1F3838A6-3556-4665-9EDF-D2C43FAFEA2A}.Release|Any CPU.Build.0 = Release|Any CPU
-		{1D76A876-71F7-4DB4-A8DB-B08D172A7F0C}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{1D76A876-71F7-4DB4-A8DB-B08D172A7F0C}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{1D76A876-71F7-4DB4-A8DB-B08D172A7F0C}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{1D76A876-71F7-4DB4-A8DB-B08D172A7F0C}.Release|Any CPU.Build.0 = Release|Any CPU
-		{E5B1B762-2478-4CAC-85E7-B18DBD05E361}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{E5B1B762-2478-4CAC-85E7-B18DBD05E361}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{E5B1B762-2478-4CAC-85E7-B18DBD05E361}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{E5B1B762-2478-4CAC-85E7-B18DBD05E361}.Release|Any CPU.Build.0 = Release|Any CPU
-		{9C589DDD-0330-42D4-AAFE-CA6779BB87F8}.Debug|Any CPU.ActiveCfg = Debug|Any CPU
-		{9C589DDD-0330-42D4-AAFE-CA6779BB87F8}.Debug|Any CPU.Build.0 = Debug|Any CPU
-		{9C589DDD-0330-42D4-AAFE-CA6779BB87F8}.Release|Any CPU.ActiveCfg = Release|Any CPU
-		{9C589DDD-0330-42D4-AAFE-CA6779BB87F8}.Release|Any CPU.Build.0 = Release|Any CPU
-	EndGlobalSection
-	GlobalSection(SolutionProperties) = preSolution
-		HideSolutionNode = FALSE
-	EndGlobalSection
-	GlobalSection(NestedProjects) = preSolution
-		{1F3838A6-3556-4665-9EDF-D2C43FAFEA2A} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
-		{1D76A876-71F7-4DB4-A8DB-B08D172A7F0C} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
-		{E5B1B762-2478-4CAC-85E7-B18DBD05E361} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
-		{9C589DDD-0330-42D4-AAFE-CA6779BB87F8} = {02EA681E-C7D8-13C7-8484-4AC65E1B71E8}
-	EndGlobalSection
-	GlobalSection(ExtensibilityGlobals) = postSolution
-		SolutionGuid = {18C5F0FE-A39D-4462-B974-4A269618CF54}
-	EndGlobalSection
-EndGlobal
