--- PromoManagementApp/PromoManagemntApp.Application/ApplicationAssemblyReference.cs
@@ -1,4 +1,0 @@
-﻿namespace PromoManagemntApp.Application
-{
-    public class ApplicationAssemblyReference { }
-}
