--- PromoManagementApp/PromoManagemntApp.Application/DTOs/Login/LoginUserDTO.cs
@@ -1,4 +1,0 @@
-﻿namespace PromoManagemntApp.Application.DTOs.Login
-{
-    public record LoginUserDTO(string Email, string Password);
-}
