--- PromoManagementApp/PromoManagemntApp.Application/DTOs/Register/RegisterUserDTO.cs
@@ -1,5 +1,0 @@
-﻿namespace PromoManagemntApp.Application.DTOs.Register
-{
-    public record RegisterUserDTO(string Name, string Lastname, string Email, 
-        string Password, string ConfirmPassword);
-}
