--- PromoManagementApp/PromoManagemntApp.Application/Implementations/RoleValidationService.cs
@@ -1,10 +1,0 @@
-﻿using PromoManagemntApp.Domain.Abstract.Services;
-using PromoManagemntApp.Domain.Constants;
-
-namespace PromoManagemntApp.Application.Implementations
-{
-    public class RoleValidationService : IRoleValidationService
-    {
-        public bool IsRoleAllowed(string role) => UserRolesConstants.AllowedRoles.Contains(role);
-    }
-}
