--- PromoManagementApp/PromoManagemntApp.Domain/Abstract/Repositories/ICampaignRepository.cs
@@ -1,6 +1,0 @@
-﻿namespace PromoManagemntApp.Domain.Abstract.Repository
-{
-    public interface ICampaignRepository
-    {
-    }
-}
