--- PromoManagementApp/PromoManagemntApp.Domain/Abstract/Services/IRoleValidationService.cs
@@ -1,7 +1,0 @@
-﻿namespace PromoManagemntApp.Domain.Abstract.Services
-{
-    public interface IRoleValidationService
-    {
-        bool IsRoleAllowed(string role);
-    }
-}
