--- PromoManagementApp/PromoManagemntApp.Domain/Constants/CampaignStatusEnum.cs
@@ -1,8 +1,0 @@
-﻿namespace PromoManagemntApp.Domain.Constants
-{
-    public enum CampaignStatusEnum
-    {
-        Active,
-        Ended
-    }
-}
