--- PromoManagementApp/PromoManagemntApp.Domain/Constants/CampaignTypeEnum.cs
@@ -1,9 +1,0 @@
-﻿namespace PromoManagemntApp.Domain.Constants
-{
-    public enum CampaignTypeEnum
-    {
-        Cashback,
-        Discount,
-        Gift
-    }
-}
