--- PromoManagementApp/PromoManagemntApp.Domain/Entities/ApplicationUser.cs
@@ -1,19 +1,0 @@
-﻿using Microsoft.AspNetCore.Identity;
-using System.ComponentModel.DataAnnotations;
-
-namespace PromoManagemntApp.Domain.Entities
-{
-    public class ApplicationUser : IdentityUser
-    {
-        [Required]
-        [MaxLength(50)]
-        public string FirstName { get; set; }
-
-        [Required]
-        [MaxLength(50)]
-        public string LastName { get; set; }
-
-        public List<Campaign>? Campaigns { get; set; }
-
-    }
-}
