--- PromoManagementApp/PromoManagemntApp.Presentation/appsettings.json
@@ -1,12 +1,0 @@
-{
-  "Logging": {
-    "LogLevel": {
-      "Default": "Information",
-      "Microsoft.AspNetCore": "Warning"
-    }
-  },
-  "AllowedHosts": "*",
-  "ConnectionStrings": {
-    "Database": "User ID=postgres;Password=qwerty1234;Host=localhost;Port=5432;Database=postgres;Pooling=true;"
-  }
-}
