+++ PromoManagementPlatform/PromoManagementPlatform.Application/ApplicationAssemblyReference.cs
@@ -1,0 +1,4 @@
+﻿namespace PromoManagementPlatform.Application
+{
+    public class ApplicationAssemblyReference { }
+}
