+++ PromoManagementPlatform/PromoManagementPlatform.Application/DTOs/Login/LoginUserDTO.cs
@@ -1,0 +1,4 @@
+﻿namespace PromoManagementPlatform.Application.DTOs.Login
+{
+    public record LoginUserDTO(string Email, string Password);
+}
