+++ PromoManagementPlatform/PromoManagementPlatform.Application/DTOs/Register/RegisterUserDTO.cs
@@ -1,0 +1,5 @@
+﻿namespace PromoManagementPlatform.Application.DTOs.Register
+{
+    public record RegisterUserDTO(string Name, string Lastname, string Email,
+        string Password, string ConfirmPassword);
+}
