+++ PromoManagementPlatform/PromoManagementPlatform.Application/Mappings/MappingProfile.cs
@@ -1,0 +1,15 @@
+﻿using AutoMapper;
+using PromoManagementPlatform.Application.DTOs.Register;
+using PromoManagementPlatform.Domain.Entities;
+
+namespace PromoManagementPlatform.Application.Mappings
+{
+    public class MappingProfile : Profile
+    {
+        public MappingProfile()
+        {
+            CreateMap<RegisterUserDTO, ApplicationUser>()
+                .ForMember(dest => dest.UserName, opt => opt.MapFrom(src => src.Email));
+        }
+    }
+}
