+++ PromoManagementPlatform/PromoManagementPlatform.Domain/Abstract/ICampaignRepository.cs
@@ -1,0 +1,6 @@
+﻿namespace PromoManagementPlatform.Domain.Abstract
+{
+    public interface ICampaignRepository
+    {
+    }
+}
