+++ PromoManagementPlatform/PromoManagementPlatform.Domain/Constants/CampaignStatusEnum.cs
@@ -1,0 +1,8 @@
+﻿namespace PromoManagementPlatform.Domain.Constants
+{
+    public enum CampaignStatusEnum
+    {
+        Active,
+        Ended
+    }
+}
