+++ PromoManagementPlatform/PromoManagementPlatform.Domain/Constants/CampaignTypeEnum.cs
@@ -1,0 +1,9 @@
+﻿namespace PromoManagementPlatform.Domain.Constants
+{
+    public enum CampaignTypeEnum
+    {
+        Cashback,
+        Discount,
+        Gift
+    }
+}
