+++ PromoManagementPlatform/PromoManagementPlatform.Domain/Constants/UserRolesConstants.cs
@@ -1,0 +1,13 @@
+﻿namespace PromoManagementPlatform.Domain.Constants
+{
+    public static class UserRolesConstants
+    {
+        public const string Admin = nameof(Admin);
+
+        public const string Manager = nameof(Manager);
+
+        public const string FinanceSpecialist = nameof(FinanceSpecialist);
+
+        public const string Unverified = nameof(Unverified);
+    }
+}
