+++ PromoManagementPlatform/PromoManagementPlatform.Domain/Entities/ApplicationUser.cs
@@ -1,0 +1,19 @@
+﻿using Microsoft.AspNetCore.Identity;
+using System.ComponentModel.DataAnnotations;
+
+namespace PromoManagementPlatform.Domain.Entities
+{
+    public class ApplicationUser : IdentityUser
+    {
+        [Required]
+        [MaxLength(50)]
+        public string FirstName { get; set; }
+
+        [Required]
+        [MaxLength(50)]
+        public string LastName { get; set; }
+
+        public List<Campaign>? Campaigns { get; set; }
+
+    }
+}
